Crate dpsim_api

DPSIM Service Rest API for controlling DPSim analyzer

Author : Richard Marston rmarston@eonerc.rwth-aachen.de

Table of endpoints

Endpoint	Method	Description	Implementation	Parameters	Returns

/simulation	POST	Add a simulation	post_simulation()	SimulationForm	Simulation

/simulation	GET	List simulations	get_simulation()	None	[ Simulation ]

/simulation/ [id]	GET	Simulation details	todo!	None	Simulation

/simulation/ [id] /results	GET	Simulation results	todo!	None	plain text

/simulation/ [id] /logs	GET	Simulation logs	todo!	None	plain text

/debug	GET	DPsim-api debug	todo!	None	plain text

Structs

Simulation

SimulationForm
Form for submitting a new Simulation

Enums

SimulationType

Functions

get_simulation
List the simulations

incomplete_form

main
The main entry point for Rocket

post_simulation
Create a new simulation

read_u64
Utility function for reading an int value from a key in a Redis DB

write_u64
Utility function for writing an int value to a key in a Redis DB

Type Definitions

SimulationJson