Module dpsim_api::routes

Structs

Route 🔒

A struct used for sharing info about a route with a template

RoutesContext 🔒

A struct used for sharing info about some routes with a template

Simulation

Struct for encapsulation Simulation details

SimulationArray

Struct for encapsulation Simulation details

SimulationError

SimulationForm

Form for submitting a new Simulation

SimulationSummary

Enums

DomainType

Enum for the various Simulation types

SimulationType

Enum for the various Simulation types

SolverType

Enum for the various Solver types

Functions

document_link 🔒

Create a link to the documentation page for the given function

get_api

List the endpoints

get_root

Redirects to /api

get_routes

Returns the list of routes that we have defined

get_simulation_id

Get the details for a simulation

get_simulations

List the simulations

incomplete_form

Handler for when an incomplete form has been submitted

okapi_add_operation_for_get_api_

okapi_add_operation_for_get_root_

okapi_add_operation_for_get_simulation_id_

okapi_add_operation_for_get_simulations_

okapi_add_operation_for_post_simulation_

parse_simulation_form 🔒

post_simulation

Create a new simulation

Type Definitions

SimulationResult 🔒