Module dpsim_api::routes

Macros

create_endpoint_with_doc

Structs

Route

A struct used for sharing info about a route with a template

RoutesContext

A struct used for sharing info about some routes with a template

SimulationError

SimulationFormStruct

Form for submitting a new Simulation

Functions

document_link

Create a link to the documentation page for the given function

document_link_page

Create an html button linking to the documentation page for the given function

get_api

List the endpoints

get_post_simulation

get_root

Redirects to /api

get_routes

Returns the list of routes that we have defined

get_simulation_id

Show details for a simulation

get_simulations

List the simulations

incomplete_form

Handler for when an incomplete form has been submitted

okapi_add_operation_for_get_api_

okapi_add_operation_for_get_post_simulation_

okapi_add_operation_for_get_root_

okapi_add_operation_for_get_simulation_id_

okapi_add_operation_for_get_simulations_

okapi_add_operation_for_post_simulation_

okapi_add_operation_for_swagger_endpoint_

osstr_to_string

parse_simulation_form

post_simulation

Create a new simulation

read_zip

Function to read a zip file

swagger_endpoint

Type Definitions

SimulationResult